git-svn-id: svn://192.168.1.10/public/CoffeeMud@8262 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Libraries/CommonMsgs.java
@@ -1067,7 +1067,7 @@ public class CommonMsgs extends StdLibrary implements CommonCommands
 			else
 			{
 				for(int i=0;i<continues.size()-1;i++)
-					str.append(Directions.getDirectionName(continues.elementAt(i)).toLowerCase().trim()+", ");
+					str.append(Directions.getDirectionName(continues.elementAt(i).intValue()).toLowerCase().trim()+", ");
 				str.append("and "+Directions.getInDirectionName(continues.lastElement().intValue()).trim()+".");
 			}
     	}
